fix(server): finish workspace API ownership
The initial extraction still retained root-owned mutable config, allowed fixtures to replace live dependencies, and relied on simultaneous root cancellation for Workspace shutdown. Those seams made the new package boundary race-prone and left its Git-heavy tests coupled to the root server binary.

Publish immutable committed snapshots, give Workspace an explicit bounded lifecycle below later-started consumers, and construct dependencies before Start. Move the associated handler and wire coverage into the Workspace lanes so private test controls and compatibility aliases are no longer part of the production contract.
