+++ internal/server/workspaceapi/lifecycle.go
@@ -9,6 +9,94 @@ import (
 	"go.kenn.io/middleman/internal/workspace/localruntime"
 )
 
+// Start launches workspace-owned background work once. The parent context is
+// observed in addition to explicit Shutdown calls so root cancellation cannot
+// strand the domain.
+func (h *Handler) Start(parent context.Context, disableMonitors bool) {
+	if h == nil {
+		return
+	}
+	if parent == nil {
+		parent = context.Background()
+	}
+	h.lifecycleMu.Lock()
+	if h.lifecycleStarted || h.lifecycleStopping {
+		h.lifecycleMu.Unlock()
+		return
+	}
+	h.lifecycleStarted = true
+	h.lifecycleMu.Unlock()
+
+	h.runBackground(func(ctx context.Context) {
+		select {
+		case <-parent.Done():
+			h.stopBackground()
+		case <-ctx.Done():
+		}
+	})
+	if h.workspaces != nil && !disableMonitors {
+		h.runBackground(h.runWorkspacePRMonitorLoop)
+		h.runBackground(h.runWorkspacePushedHeadObserverLoop)
+	}
+}
+
+func (h *Handler) runBackground(run func(context.Context)) bool {
+	if h == nil || run == nil {
+		return false
+	}
+	h.lifecycleMu.Lock()
+	if h.lifecycleStopping {
+		h.lifecycleMu.Unlock()
+		return false
+	}
+	h.lifecycleWG.Add(1)
+	ctx := h.lifecycleCtx
+	h.lifecycleMu.Unlock()
+	go func() {
+		defer h.lifecycleWG.Done()
+		run(ctx)
+	}()
+	return true
+}
+
+func (h *Handler) stopBackground() {
+	h.lifecycleMu.Lock()
+	if h.lifecycleStopping {
+		h.lifecycleMu.Unlock()
+		return
+	}
+	h.lifecycleStopping = true
+	h.lifecycleCancel()
+	done := h.lifecycleDone
+	h.lifecycleMu.Unlock()
+	go func() {
+		h.lifecycleWG.Wait()
+		if h.workspaceDiffCache != nil {
+			h.workspaceDiffCache.Wait()
+		}
+		close(done)
+	}()
+}
+
+// Shutdown stops Workspace background work and waits for it within ctx. It is
+// safe to call concurrently and repeatedly; later callers can wait with a
+// longer context after an earlier deadline expires.
+func (h *Handler) Shutdown(ctx context.Context) error {
+	if h == nil {
+		return nil
+	}
+	if ctx == nil {
+		ctx = context.Background()
+	}
+	h.stopBackground()
+	select {
+	case <-h.lifecycleDone:
+		return nil
+	case <-ctx.Done():
+		return ctx.Err()
+	}
+}
+
 // RestoreRuntimeSessions restores persisted workspace runtime sessions after
 // the runtime manager has been constructed.
 func (h *Handler) RestoreRuntimeSessions(ctx context.Context) error {
@@ -70,9 +158,6 @@ func (h *Handler) HandleRuntimeSessionExit(info localruntime.SessionInfo) {
 		return
 	}
 	h.invalidateWorkspaceEnrichment(info.WorkspaceID)
-	if h.runBackground == nil {
-		return
-	}
 	h.runBackground(func(ctx context.Context) {
 		cleanupCtx, cancel := context.WithTimeout(ctx, 2*time.Second)
 		defer cancel()
