+++ internal/server/workspaceapi/workspace_branch_actions_test.go
@@ -1,0 +1,38 @@
+package workspaceapi
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"go.kenn.io/middleman/internal/db"
+	"go.kenn.io/middleman/internal/testutil/dbtest"
+	"go.kenn.io/middleman/internal/workspace"
+)
+
+func TestRevealWorkspaceOpensWorkspacePath(t *testing.T) {
+	require := require.New(t)
+	assert := assert.New(t)
+	database := dbtest.Open(t)
+	manager := workspace.NewManager(database, t.TempDir())
+	path := t.TempDir()
+	require.NoError(database.InsertWorkspace(t.Context(), &db.Workspace{
+		ID: "ws-reveal", Platform: "github", PlatformHost: "github.com",
+		RepoOwner: "acme", RepoName: "widget", ItemType: db.WorkspaceItemTypePullRequest,
+		ItemNumber: 1, WorktreePath: path, Status: "ready",
+	}))
+	h := New(Deps{DB: database, Workspaces: manager})
+	previous := revealWorkspacePath
+	t.Cleanup(func() { revealWorkspacePath = previous })
+	var opened string
+	revealWorkspacePath = func(_ context.Context, got string) error {
+		opened = got
+		return nil
+	}
+
+	_, err := h.revealWorkspace(t.Context(), &revealWorkspaceInput{ID: "ws-reveal"})
+
+	require.NoError(err)
+	assert.Equal(path, opened)
+}
