+++ internal/server/workspacetest/project_git_routes_test.go
@@ -1,4 +1,4 @@
-package server
+package workspacetest
 
 import (
 	"encoding/json"
@@ -12,20 +12,150 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	dbpkg "go.kenn.io/middleman/internal/db"
+	"go.kenn.io/middleman/internal/db"
+	"go.kenn.io/middleman/internal/procutil"
 )
 
-// TestListProjectBranches covers GET /api/v1/projects/{id}/branches: the
-// project repository's local branch names, sorted.
-func TestListProjectBranches(t *testing.T) {
+func TestCloneProject(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
 	if _, err := exec.LookPath("git"); err != nil {
 		t.Skip("git not available")
 	}
 	require := require.New(t)
 	assert := assert.New(t)
 
-	srv, _ := setupTestServer(t)
+	srv, database := setupProjectServer(t)
+	ts := httptest.NewServer(srv)
+	defer ts.Close()
+
+	source := initLifecycleRouteRepo(t)
+	runGit(t, source, "branch", "feat/clone")
+	dest := filepath.Join(t.TempDir(), "cloned")
+
+	body := mustMarshal(t, map[string]any{
+		"url":  source,
+		"path": dest,
+	})
+	resp := httpDo(t, ts, http.MethodPost, "/api/v1/projects/clone", body)
+	require.Equal(http.StatusCreated, resp.StatusCode)
+	var created struct {
+		ID        string `json:"id"`
+		LocalPath string `json:"local_path"`
+	}
+	require.NoError(json.NewDecoder(resp.Body).Decode(&created))
+	resp.Body.Close()
+	require.NotEmpty(created.ID)
+	assert.Equal(dest, created.LocalPath)
+
+	// The checkout exists on disk and the project is registered with
+	// its root worktree discovered.
+	_, err := os.Stat(filepath.Join(dest, ".git"))
+	require.NoError(err)
+	project, err := database.GetProjectByID(t.Context(), created.ID)
+	require.NoError(err)
+	assert.Equal(dest, project.LocalPath)
+	rows := listWorktreeRows(t, ts, created.ID)
+	require.NotEmpty(rows, "discovery registers the root checkout")
+
+	// Cloning onto an existing destination is refused.
+	resp = httpDo(t, ts, http.MethodPost, "/api/v1/projects/clone", body)
+	assert.Equal(http.StatusConflict, resp.StatusCode)
+	resp.Body.Close()
+}
+
+// TestCloneProjectBranchAndHomePath covers the branch option and
+// home-relative destination expansion: fleet clients send "~/..." paths
+// because only the executing host knows its home.
+func TestCloneProjectBranchAndHomePath(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
+	if _, err := exec.LookPath("git"); err != nil {
+		t.Skip("git not available")
+	}
+	require := require.New(t)
+	assert := assert.New(t)
+
+	fakeHome := t.TempDir()
+	t.Setenv("HOME", fakeHome)
+
+	srv, _ := setupProjectServer(t)
+	ts := httptest.NewServer(srv)
+	defer ts.Close()
+
+	source := initLifecycleRouteRepo(t)
+	runGit(t, source, "branch", "feat/clone")
+
+	body := mustMarshal(t, map[string]any{
+		"url":    source,
+		"path":   "~/clones/widget",
+		"branch": "feat/clone",
+	})
+	resp := httpDo(t, ts, http.MethodPost, "/api/v1/projects/clone", body)
+	require.Equal(http.StatusCreated, resp.StatusCode)
+	var created struct {
+		LocalPath string `json:"local_path"`
+	}
+	require.NoError(json.NewDecoder(resp.Body).Decode(&created))
+	resp.Body.Close()
+	assert.Equal(filepath.Join(fakeHome, "clones", "widget"), created.LocalPath)
+
+	out, err := procutil.Command(
+		"git", "-C", created.LocalPath,
+		"rev-parse", "--abbrev-ref", "HEAD",
+	).Output()
+	require.NoError(err)
+	assert.Equal("feat/clone", string(out[:len(out)-1]))
+}
+
+// TestCloneProjectFailureCleansOwnedDestination pins the rollback
+// contract: a failed clone removes the destination directory this
+// request reserved, so an immediate retry reaches git again instead of
+// tripping destinationExists over a leftover partial checkout.
+func TestCloneProjectFailureCleansOwnedDestination(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
+	if _, err := exec.LookPath("git"); err != nil {
+		t.Skip("git not available")
+	}
+	require := require.New(t)
+	assert := assert.New(t)
+
+	srv, _ := setupProjectServer(t)
+	ts := httptest.NewServer(srv)
+	defer ts.Close()
+
+	dest := filepath.Join(t.TempDir(), "clones", "broken")
+	body := mustMarshal(t, map[string]any{
+		// A file URL that does not exist: git fails after the
+		// destination has been reserved.
+		"url":  "file://" + filepath.Join(t.TempDir(), "no-such-repo.git"),
+		"path": dest,
+	})
+
+	for attempt := 1; attempt <= 2; attempt++ {
+		resp := httpDo(t, ts, http.MethodPost, "/api/v1/projects/clone", body)
+		var problem struct {
+			Code string `json:"code"`
+		}
+		require.NoError(json.NewDecoder(resp.Body).Decode(&problem))
+		resp.Body.Close()
+		require.Equal(http.StatusBadRequest, resp.StatusCode,
+			"attempt %d must fail at git, not at destination reservation", attempt)
+		assert.NotEqual("destinationExists", problem.Code,
+			"attempt %d must not trip over a leftover partial checkout", attempt)
+		_, statErr := os.Stat(dest)
+		assert.True(os.IsNotExist(statErr),
+			"attempt %d must remove the reserved destination", attempt)
+	}
+}
+
+func TestListProjectBranches(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
+	if _, err := exec.LookPath("git"); err != nil {
+		t.Skip("git not available")
+	}
+	require := require.New(t)
+	assert := assert.New(t)
+
+	srv, _ := setupProjectServer(t)
 	ts := httptest.NewServer(srv)
 	defer ts.Close()
 
@@ -54,13 +184,14 @@ func TestListProjectBranches(t *testing.T) {
 // GET /api/v1/projects/{pid}/worktrees/{wid}/inspect: dirty state, live
 // session count, and branch-delete eligibility for delete confirmation UIs.
 func TestInspectProjectWorktree(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
 	if _, err := exec.LookPath("git"); err != nil {
 		t.Skip("git not available")
 	}
 	require := require.New(t)
 	assert := assert.New(t)
 
-	srv, _ := setupTestServer(t)
+	srv, _ := setupProjectServer(t)
 	ts := httptest.NewServer(srv)
 	defer ts.Close()
 
@@ -130,13 +261,14 @@ func TestInspectProjectWorktree(t *testing.T) {
 // runtime session) still counts as alive — otherwise delete confirmation
 // under-reports live work after a restart.
 func TestInspectProjectWorktreeCountsStoredTmuxSessions(t *testing.T) {
+	acquireWorkspaceGitSlot(t)
 	if _, err := exec.LookPath("git"); err != nil {
 		t.Skip("git not available")
 	}
 	require := require.New(t)
 	assert := assert.New(t)
 
-	srv, database := setupTestServer(t)
+	srv, database := setupProjectServer(t)
 	ts := httptest.NewServer(srv)
 	defer ts.Close()
 
@@ -148,7 +280,7 @@ func TestInspectProjectWorktreeCountsStoredTmuxSessions(t *testing.T) {
 	worktreeID := root["id"].(string)
 
 	require.NoError(database.UpsertProjectWorktreeTmuxSession(
-		t.Context(), &dbpkg.ProjectWorktreeTmuxSession{
+		t.Context(), &db.ProjectWorktreeTmuxSession{
 			WorktreeID:  worktreeID,
 			SessionKey:  "restart-survivor",
 			SessionName: "middleman-restart-survivor",
@@ -173,37 +305,3 @@ func TestInspectProjectWorktreeCountsStoredTmuxSessions(t *testing.T) {
 // restart case with a runtime manager configured: the in-memory runtime
 // has no entry for the stored row (the daemon restarted), and the
 // merged listing must still count the durable session as alive.
-func TestInspectProjectWorktreeCountsStoredTmuxSessionsWithRuntime(t *testing.T) {
-	if _, err := exec.LookPath("git"); err != nil {
-		t.Skip("git not available")
-	}
-	require := require.New(t)
-	assert := assert.New(t)
-
-	srv, projectID, worktreeID, _ :=
-		setupProjectWorktreeCommandSessionTestWithRecord(t)
-	require.NotNil(srv.runtime, "fixture must configure a runtime manager")
-	ts := httptest.NewServer(srv)
-	defer ts.Close()
-
-	require.NoError(srv.db.UpsertProjectWorktreeTmuxSession(
-		t.Context(), &dbpkg.ProjectWorktreeTmuxSession{
-			WorktreeID:  worktreeID,
-			SessionKey:  "restart-survivor-runtime",
-			SessionName: "middleman-restart-survivor-runtime",
-			Label:       "Survivor",
-			CreatedAt:   time.Now().UTC(),
-		},
-	))
-
-	resp := httpDo(t, ts, http.MethodGet,
-		"/api/v1/projects/"+projectID+"/worktrees/"+worktreeID+"/inspect", nil)
-	require.Equal(http.StatusOK, resp.StatusCode)
-	var got struct {
-		AliveSessionCount int `json:"alive_session_count"`
-	}
-	require.NoError(json.NewDecoder(resp.Body).Decode(&got))
-	resp.Body.Close()
-	assert.Equal(1, got.AliveSessionCount,
-		"stored tmux sessions count as alive when the runtime has no in-memory entry")
-}
